move OrderModel out of graphql module
