+++ backend/infrahub/graphql/resolvers/ipam.py
@@ -16,9 +16,10 @@ from infrahub.core.node import Node
 from infrahub.core.protocols import BuiltinIPNamespace, BuiltinIPPrefix
 from infrahub.core.schema.generic_schema import GenericSchema
 from infrahub.exceptions import ValidationError
-from infrahub.graphql.models import OrderModel
 from infrahub.graphql.parser import extract_selection
 from infrahub.graphql.permissions import get_permissions
+
+from ..order import deserialize_order_input
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -322,7 +323,7 @@ async def ipam_paginated_list_resolver(  # noqa: PLR0915
     if not isinstance(schema, GenericSchema) or schema.kind not in [InfrahubKind.IPADDRESS, InfrahubKind.IPPREFIX]:
         raise ValidationError(f"{schema.kind} is not {InfrahubKind.IPADDRESS} or {InfrahubKind.IPPREFIX}")
 
-    order_model = OrderModel.from_input(input_data=order)
+    order_model = deserialize_order_input(input_data=order)
     fields = await extract_selection(info=info, schema=schema)
     resolve_available = bool(kwargs.pop("include_available", False))
     kinds_to_filter: list[str] = kwargs.pop("kinds", [])  # type: ignore[assignment]
